requirements.txt
marker-pdf>=0.3.2
torch>=2.0.0